src/main/java/lt/dualpair/server/service/user/CustomDetailsServiceImpl.java
package lt.dualpair.server.service.user;

import lt.dualpair.server.domain.model.user.User;
import lt.dualpair.server.infrastructure.persistence.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.social.security.SocialUserDetailsService;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class CustomDetailsServiceImpl implements UserDetailsService, SocialUserDetailsService {

    private UserRepository userRepository;

    @Override
    public User loadUserByUsername(String username) throws UsernameNotFoundException {
        return findByUsername(username);
    }

    @Deprecated
    public User loadUserByUserId(String username) throws UsernameNotFoundException {
        return findByUsername(username);
    }

    private User findByUsername(String username) {
        Optional<User> user = userRepository.findByUsername(username);
        if (!user.isPresent()) {
            throw new UsernameNotFoundException("User with username " + username + " not found.");
        }
        return user.get();
    }

    @Autowired
    public void setUserRepository(UserRepository userRepository) {
        this.userRepository = userRepository;
    }
}
